Prepare for sending requests, make discovery synchronous
Prepare the command line parser for sending out a basic request, but do
not implement it yet.
At the same time, make the discovery synchronous, as we will need to
make sending out requests and handling the responses synchronous as
well, rendering the use of tokio redundant. Also, this means we can cut
down on dependencies.