Prepare for Modbus, rework AA55 protocol for identification
It seems that using the AA55 protocol is not the best choice for trying
to get the inverter data, instead the Modbus protocol should be used.

AA55 is reduced to querying the serial number and firmware version for
now, to help identify inverters (the type is encoded into the serial
number). With the information on which type of inverter is there, we can
then decide which Modbus registers to use.

Currently the only planned support is for ET-series inverters, but I
want to keep the possibility to support further inverter types in the
future.