"A few changes"
Make massive changes to the code, add all Modbus definitions and make
them printable on the console.

The Modbus part is now finished, what is remaining is to make some
changes to the metric definitions for better naming and presentation to
Prometheus, as well as the actual Prometheus statistics serving part.